+++ packages/react-core/src/components/Form/__tests__/Form.test.tsx
@@ -12,4 +12,9 @@ describe('Form component', () => {
     const view = shallow(<Form isHorizontal />);
     expect(view).toMatchSnapshot();
   });
+
+  test('should render form with limited width', () => {
+    const view = shallow(<Form isWidthLimited />);
+    expect(view).toMatchSnapshot();
+  });
 });