+++ packages/react-integration/cypress/integration/modal.spec.ts
@@ -188,4 +188,24 @@ describe('Modal Test', () => {
         });
     });
   });
+
+  it('Verify Help Modal', () => {
+    cy.get('#showHelpModalButton').then((modalButton: JQuery<HTMLButtonElement>) => {
+      cy.wrap(modalButton).click();
+      cy.get('.pf-c-modal-box__header').should('have.class', 'pf-m-help');
+      cy.get('.pf-c-modal-box')
+        .then(() => {
+          cy.get('.pf-c-modal-box .pf-c-button[aria-label="Close"]').then(closeButton => {
+            cy.wrap(closeButton).click();
+            cy.get('.pf-c-modal-box').should('not.exist');
+          });
+        })
+        .then(() => {
+          cy.wrap(modalButton).click();
+          cy.get('.pf-c-modal-box').should('exist');
+          cy.get('body').trigger('keydown', { keyCode: 27, which: 27 });
+          cy.get('.pf-c-modal-box').should('not.exist');
+        });
+    });
+  });
 });