+++ packages/react-table/src/components/Table/utils/transformers.tsx
@@ -9,6 +9,7 @@ export { collapsible, expandedRow, expandable } from './decorators/collapsible';
 export { compoundExpand } from './decorators/compoundExpand';
 export { headerCol } from './decorators/headerCol';
 export { classNames, Visibility } from './decorators/classNames';
+export { info } from './decorators/info';
 
 import { IFormatterValueType, IExtra, ITransform } from '../Table';