+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.15...@patternfly/react-icons@4.7.16) (2020-10-23)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
+## [4.7.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.14...@patternfly/react-icons@4.7.15) (2020-10-22)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
+## [4.7.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.13...@patternfly/react-icons@4.7.14) (2020-10-22)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
 ## [4.7.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.12...@patternfly/react-icons@4.7.13) (2020-10-20)
 
 **Note:** Version bump only for package @patternfly/react-icons