+++ packages/react-integration/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.58.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.58.0...@patternfly/react-integration@4.58.1) (2020-10-26)
+
+
+### Bug Fixes
+
+* **Alert, NotificationDrawer:** Added tooltips to truncated titles ([#5043](https://github.com/patternfly/patternfly-react/issues/5043)) ([eb77e52](https://github.com/patternfly/patternfly-react/commit/eb77e52337734f462f2ce5f5fd51fbbf3c4aef0e))
+
+
+
+
+
+# [4.58.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.57.0...@patternfly/react-integration@4.58.0) (2020-10-23)
+
+
+### Features
+
+* **Pagination:** add sticky modifier flag ([#4977](https://github.com/patternfly/patternfly-react/issues/4977)) ([447772a](https://github.com/patternfly/patternfly-react/commit/447772a2ae1edd3db43fb8088377d72e10c77ade))
+
+
+
+
+
+# [4.57.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.56.0...@patternfly/react-integration@4.57.0) (2020-10-23)
+
+
+### Features
+
+* **Splitter:** add Splitter component ([#5051](https://github.com/patternfly/patternfly-react/issues/5051)) ([70073f2](https://github.com/patternfly/patternfly-react/commit/70073f27cf7833ebf4785887636ba6474f53de8b))
+
+
+
+
+
+# [4.56.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.55.0...@patternfly/react-integration@4.56.0) (2020-10-22)
+
+
+### Features
+
+* **Modal:** add help section to the modal header ([#5002](https://github.com/patternfly/patternfly-react/issues/5002)) ([6a2b79b](https://github.com/patternfly/patternfly-react/commit/6a2b79bf3b79652eeb11f89cb47225deeaa8aaee))
+
+
+
+
+
+# [4.55.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.54.0...@patternfly/react-integration@4.55.0) (2020-10-22)
+
+
+### Features
+
+* **ContextSelector:** add footer to selector ([#4976](https://github.com/patternfly/patternfly-react/issues/4976)) ([e2738d9](https://github.com/patternfly/patternfly-react/commit/e2738d9aa298787f238cebdaa64ac52f1c09b5fa))
+
+
+
+
+
 # [4.54.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@4.53.0...@patternfly/react-integration@4.54.0) (2020-10-20)